SELFOUR-29: install fault endpoint into tcb cnode
This change installs the fault endpoint cap into the tcb cnode
on seL4_TCB_Configure or seL4_TCB_SetSpace, first validating it.

This means either of those functions will now return an error
 if the cap is not either a null cap or an endpoint with send and
grant rights.

Significantly, the cap passed to the function should be in the callers
cspace, not the target tcbs.